// packages/sdk/src/generated/leash/types/agentStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

/** Whether an agent may pay. */
export const AgentStatus = {
  0: "Active",
  1: "Frozen",
  Active: 0,
  Frozen: 1,
} as const;

/** Whether an agent may pay. */
export type AgentStatus = (typeof AgentStatus)[Exclude<
  keyof typeof AgentStatus,
  number
>];

export type AgentStatusArgs = AgentStatus;

export function getAgentStatusEncoder(): FixedSizeEncoder<AgentStatusArgs> {
  return getEnumEncoder(AgentStatus);
}

export function getAgentStatusDecoder(): FixedSizeDecoder<AgentStatus> {
  return getEnumDecoder(AgentStatus as Omit<typeof AgentStatus, number>);
}

export function getAgentStatusCodec(): FixedSizeCodec<
  AgentStatusArgs,
  AgentStatus
> {
  return combineCodec(getAgentStatusEncoder(), getAgentStatusDecoder());
}